# docs/api/swagger.yaml
components:
  schemas:
...
    ApiError:
      properties:
        code:
          example: BAD_REQUEST
          type: string
        details:
          description: |-
            Details contains additional error context (field names, validation rules, etc.)
            Type varies by error code: map[string]string for validation, string for general errors
          example: '{"field":"email","rule":"required"}'
        message:
          example: Invalid request parameters
          type: string
      type: object
...
    CreateRepository:
      properties:
        config:
          $ref: '#/components/schemas/RepoConfig'
        name:
          example: owner/repo
          type: string
      required:
      - name
      type: object
...
    EffectiveConfig:
      properties:
        effective:
          $ref: '#/components/schemas/RepoConfig'
        enabled:
          type: boolean
        globalConfig:
          $ref: '#/components/schemas/RepoConfig'
        hasRepoConfig:
          type: boolean
        repoConfig:
          $ref: '#/components/schemas/RepoConfig'
        repositoryId:
          type: string
        repositoryName:
          example: owner/repo
          type: string
      type: object
...
    PaginatedResponse-Repository:
      properties:
        data:
          items:
            $ref: '#/components/schemas/Repository'
          type: array
          uniqueItems: false
        pagination:
          $ref: '#/components/schemas/Pagination'
      type: object
...
    Pagination:
      properties:
        page:
          example: 1
          type: integer
        pageSize:
          example: 20
          type: integer
        totalItems:
          example: 100
          type: integer
        totalPages:
          example: 5
          type: integer
      type: object
...
    RepoConfig:
      properties:
        complexityThreshold:
          example: 70
          type: integer
        defaultModel:
          example: claude-3
          type: string
        forceDesignConfirm:
          example: false
          type: boolean
        maxConcurrency:
          example: 5
          type: integer
        taskRetryLimit:
          example: 3
          type: integer
      type: object
    Repository:
      properties:
        config:
          $ref: '#/components/schemas/RepoConfig'
        enabled:
          example: true
          type: boolean
        id:
          example: 550e8400-e29b-41d4-a716-446655440000
          type: string
        name:
          example: owner/repo
          type: string
      type: object
...
    Success:
      properties:
        message:
          example: Operation completed
          type: string
        status:
          example: success
          type: string
      type: object
...
    UpdateRepository:
      properties:
        config:
          $ref: '#/components/schemas/RepoConfig'
      type: object
...
info:
  description: Automated development agent system - from GitHub Issue to Pull Request
  title: Litchi API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/repositories:
    get:
      description: Get all repository configurations with optional pagination and
        filtering
      parameters:
      - description: Page number (1-based)
        in: query
        name: page
        schema:
          default: 1
          type: integer
      - description: Items per page
        in: query
        name: pageSize
        schema:
          default: 20
          type: integer
      - description: Filter by enabled status (true, false, all)
        in: query
        name: enabled
        schema:
          default: all
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse-Repository'
          description: List of repositories
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Internal server error
      summary: List repositories
      tags:
      - repositories
    post:
      description: Create a new repository configuration for processing
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/CreateRepository'
                description: Repository creation request
                summary: body
        description: Repository creation request
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Repository'
          description: Created repository
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Invalid request body
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Repository already exists
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Internal server error
      summary: Create repository
      tags:
      - repositories
  /api/v1/repositories/{name}:
    delete:
      description: Delete repository configuration by name
      parameters:
      - description: Repository name (owner/repo)
        in: path
        name: name
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
          description: Repository deleted
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Repository not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Internal server error
      summary: Delete repository
      tags:
      - repositories
    get:
      description: Get repository configuration by name (owner/repo format)
      parameters:
      - description: Repository name (owner/repo)
        in: path
        name: name
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Repository'
          description: Repository configuration
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Repository not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Internal server error
      summary: Get repository
      tags:
      - repositories
    put:
      description: Update repository configuration by name
      parameters:
      - description: Repository name (owner/repo)
        in: path
        name: name
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/UpdateRepository'
                description: Repository update request
                summary: body
        description: Repository update request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Repository'
          description: Updated repository
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Invalid request body
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Repository not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Internal server error
      summary: Update repository
      tags:
      - repositories
  /api/v1/repositories/{name}/disable:
    post:
      description: Disable a repository from processing incoming GitHub events
      parameters:
      - description: Repository name (owner/repo)
        in: path
        name: name
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
          description: Repository disabled
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Repository not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Internal server error
      summary: Disable repository
      tags:
      - repositories
  /api/v1/repositories/{name}/effective-config:
    get:
      description: Get the merged effective configuration (global + repository overrides)
      parameters:
      - description: Repository name (owner/repo)
        in: path
        name: name
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EffectiveConfig'
          description: Effective configuration
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Internal server error
      summary: Get effective config
      tags:
      - repositories
  /api/v1/repositories/{name}/enable:
    post:
      description: Enable a repository to process incoming GitHub events
      parameters:
      - description: Repository name (owner/repo)
        in: path
        name: name
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
          description: Repository enabled
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Repository not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Internal server error
      summary: Enable repository
      tags:
      - repositories
...